+++ sandbox/Untitled-1
@@ -6,14 +6,13 @@ class TrackList
 class TrackListSpotify < SpotifyConnector, TrackList
 class TrackListBillboards < BillboardsConnector, TrackList
 
-class Featurizer 
+class Featurizer (DONE)
     provide audio_db_path 
     provide track name
     saves the features out to tracks
     loads the features up from tracks
 
-
-class FeatureSpotifyLow < Featurizer
+class FeatureSpotifyLow < Featurizer (DONE)
     sets feature name as 'spotify_low_audio'
     sets features included as 'spotify_audio_extra'
     calculates relevant features for input track
@@ -22,8 +21,7 @@ class FeatureSpotifyLow < Featurizer
     loads relevant features for input tracks
 
 class FeatureMelSpectrogram < Featurizer
-class FeatureMusicnn < Featurizer
-class FeatureMusicnn < Featurizer
+class FeatureMusicnn < Featurizer (DONE)
 
 class Recommender 
     recommender_maker
