--- src/Featurizers/SpotifyMusicalityFeaturizer.py
@@ -1,4 +1,0 @@
-from .Featurizer import Featurizer
-
-class SpotifyMusicalityFeaturizer(Featurizer):
-    pass
