+++ src/TrackLists/__init__.py
@@ -1,0 +1,2 @@
+from TrackList import *
+from TrackListSpotify import *
