+++ src/__init__.py
@@ -1,5 +1,6 @@
 from .SpotifyConnector import SpotifyConnector
 from .DatabaseSpotifyAudio import DatabaseSpotifyAudio
 from .SpotifyUserExplorer import SpotifyUserExplorer
-from .TrackList import TrackListSpotify, TrackList
 from .Featurizers import *
+from .TrackLists import *
+from .Recommenders import *
